% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{appendToDropzone}
\alias{appendToDropzone}
\title{Append item to end of specified dropzone}
\usage{
appendToDropzone(session, value, dragzoneId, dropzoneId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{value}{The value to clone from dragzoneId and append to dropzoneId.}

\item{dragzoneId}{The \code{id} of the dragzone.}

\item{dropzoneId}{The \code{id} of the dropzone.}
}
\description{
Append item to end of specified dropzone
}
