% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\docType{package}
\name{dndselectr}
\alias{dndselectr}
\alias{dndselectr-package}
\title{dndselectr: Implements a drag-and-drop Shiny select input}
\description{
Implements a drag-and-drop Shiny select input. This implementation creates a Shiny input that replicates
much of the functionality of selectInput. Multiple zones for dragging and dropping are allowed.
Currently utilizes Dragula JS library, https://github.com/bevacqua/dragula.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

# basic example
shinyApp(
  ui = fluidPage(
    column(6,
      dragZone("dragzone", choices = list(one = "One",
                                          two = "Two",
                                          three = "Three",
                                          four = "Four"))
      ),
    column(6,
      dropZoneInput("dropzone", choices = list(one = "1",
                                               two = "2",
                                               three = "3",
                                               four = "4"),
                                highlight = TRUE)
      )
  ),
  server = function(input, output) {
  }
)
}

}
\seealso{
\code{\link{dragZone}}, \code{\link{dropZoneInput}}
}
