% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{dragulaZoneItems}
\alias{dragulaZoneItems}
\title{Create items for dragzones/dropzones}
\usage{
dragulaZoneItems(zone, type, items, ids = rep(NA, length(items)),
  selected = NULL)
}
\arguments{
\item{zone}{Container zone type: either \code{drop} or \code{drag}}

\item{type}{Are these \code{options} or \code{presets}?}

\item{items}{List of item labels, with names corresponding to values}

\item{ids}{If multivalued, these will be unique ids}

\item{selected}{Unique values (of form <value>-ds-<id> if multivalued).}
}
\value{
div element
}
\description{
In Dragula, the drag and drop areas are called "containers". This function
creates the individual draggable items and options in these containers.
}
