% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{dragulaZoneOptions}
\alias{dragulaZoneOptions}
\title{Create options for dragzones/dropzones}
\usage{
dragulaZoneOptions(type, value, label = NULL)
}
\arguments{
\item{type}{Container type: either \code{drop} or \code{drag}}

\item{value}{Value for allowable item/option (unique identifier)}

\item{label}{Label for allowable item/option (what the user sees)}
}
\value{
div element
}
\description{
In Dragula, the drag and drop areas are called "containers". This function
creates the individual draggable items in these containers.
}
