% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{dragZone}
\alias{dragZone}
\title{Create a Dragula dragzone container}
\usage{
dragZone(id, choices)
}
\arguments{
\item{id}{The container id.}

\item{choices}{List of values to select from.}
}
\description{
Create a Dragula dragzone container
}
\examples{
dragZone("dragzone", choices = list(one = "One",
                                    two = "Two",
                                    three = "Three",
                                    four = "Four"))

}
\seealso{
\code{\link{dragulaSelectR}}
}
