% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{dropZoneServer}
\alias{dropZoneServer}
\title{Create server code to handle server-side creation of UI for dropzone items}
\usage{
dropZoneServer(session, dropZoneId, server)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{dropZoneId}{The \code{id} of the first dropzone.}

\item{server}{Function or function name as a string that will be used for
server-side creation of UI for dropzone items. This is needed only when the
dropzone items contain Shiny inputs and/or outputs.}
}
\value{
Expression including an observe event handler.
}
\description{
This function will create the necessary event handlers that will call
  the UI function specified in the \code{server} argument of the
  \code{\link{dropZoneInput}} function.
}
