% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{entangle}
\alias{entangle}
\title{Entangle two dropzones}
\usage{
entangle(session, dropZoneOneId, dropZoneTwoId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{dropZoneOneId}{The \code{id} of the first dropzone.}

\item{dropZoneTwoId}{The \code{id} of the second dropzone.}
}
\value{
Expression including the two observe events.
}
\description{
Create observe events that entangle two Shiny dropzones. Useful for
hidden dropzones that take drops but display options elsewhere.
}
