% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{entangleInputs}
\alias{entangleInputs}
\title{Entangle inputs}
\usage{
entangleInputs(session, ...)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{...}{The \code{ids} of the inputs}
}
\value{
Expressions including the \code{observeEvent}s.
}
\description{
Create observe events that entangle multiple Shiny inputs. Useful for
hidden dropzones that take drops but display options elsewhere.
}
