% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{entangleSourceToTarget}
\alias{entangleSourceToTarget}
\title{Update target dropzone entangled with source dropzone}
\usage{
entangleSourceToTarget(session, sourceId, targetId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{sourceId}{The \code{id} of the source dropzone.}

\item{targetId}{The \code{id} of the target dropzone.}
}
\description{
Update target dropzone entangled with source dropzone
}
