% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{multivalues}
\alias{multivalues}
\title{Convert unique values to multivalues}
\usage{
multivalues(values, ids = FALSE)
}
\arguments{
\item{ids}{Return unique ids rather than multivalues?}

\item{uniqueVals}{Values returned by dropzone input}
}
\value{
Multivalues with unique id stripped away.
}
\description{
(Multivalued dropzones only) This will drop the added unique counter ids
and convert to multivalued inputs.
}
