% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{opts2class}
\alias{opts2class}
\title{Converts options to class names}
\usage{
opts2class(varArgs)
}
\arguments{
\item{varArgs}{Named list of options for dropzone}
}
\value{
String
}
\description{
Converts options to class names
}
