% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{removeSelected}
\alias{removeSelected}
\title{Delete selected item in specified dropzone}
\usage{
removeSelected(session, dropzoneId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{dropzoneId}{The \code{id} of the dropzone.}
}
\description{
Delete selected item in specified dropzone
}
