% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{select}
\alias{select}
\title{Select element}
\usage{
select(session, value, dropzoneId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{value}{The value of the item to select.}

\item{dropzoneId}{The \code{id} of the dropzone.}
}
\description{
Select element
}
