% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{unselect}
\alias{unselect}
\title{Remove selection}
\usage{
unselect(session, dropzoneId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{dropzoneId}{The \code{id} of the dropzone.}
}
\description{
Remove selection
}
