% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{updateDragZone}
\alias{updateDragZone}
\title{Change the choices of a dragzone on the client}
\usage{
updateDragZone(session = getDefaultReactiveDomain(), id, choices = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{id}{The container id.}

\item{choices}{List of values to select from.}
}
\description{
Change the choices of a dragzone on the client
}
