% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{updateDropZoneInput}
\alias{updateDropZoneInput}
\title{Change the values or settings of a dropzone on the client}
\usage{
updateDropZoneInput(session, inputId, presets = NULL, placeholder = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The \code{input} slot that will be used to acces the value.}

\item{presets}{Array or list of preset values.}

\item{placeholder}{Insert placeholder text.}
}
\description{
Note: Set presets to NA if you want to delete all presets. Will give a warning.
}
