% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{attachDependencies}
\alias{attachDependencies}
\title{Attach javascript dependencies}
\usage{
attachDependencies(...)
}
\arguments{
\item{...}{Shiny tag object}
}
\description{
\code{attachDependencies} attaches the javascript dependencies. Specifically,
the Dragula JS package is attached, as well as the javascript wrapper and
input bindings.
}
