% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{createContainerChoices}
\alias{createContainerChoices}
\title{Create options for dragzones/dropzones}
\usage{
createContainerChoices(value, label, type)
}
\arguments{
\item{value}{Value for allowable item/option (unique identifier)}

\item{label}{Label for allowable item/option (what the user sees)}

\item{class}{One of 'ds-dropoption' or 'ds-dragitem'}
}
\value{
div element
}
\description{
In Dragula, the drag and drop areas are called "containers". This function
creates the individual draggable items in these containers.
}
