% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{dragulaZoneItems}
\alias{dragulaZoneItems}
\title{Create items for dragzones/dropzones}
\usage{
dragulaZoneItems(zone, type, items, ids = rep(NA, length(items)),
  selected = NULL, invisible = NULL, locked = NULL, freeze = NULL,
  togglevis = FALSE, togglelock = FALSE)
}
\arguments{
\item{zone}{Container zone type: either \code{drop} or \code{drag}}

\item{type}{Are these \code{options} or \code{presets}?}

\item{items}{List of item labels, with names corresponding to values}

\item{ids}{If multivalued, these will be unique ids}

\item{selected}{Selected items (array length of items - either NA or ds-selected)}

\item{invisible}{Invisible items (array length of items - either NA or ds-invisible)}

\item{locked}{Locked items (array length of items - either NA or ds-locked)}

\item{freeze}{No items allowed before these. Analogous to freezing the first few
columns of a spreadsheet (array length of items - either NA or ds-freeze). Makes
since only for first initial items, and when used in conjunction with locked.
Frozen items also cannot be toggled.}

\item{togglevis}{Add an icon to allow toggling items between visible/invisible.}

\item{togglelock}{Add an icon to allow toggling items between locked/unlocked.}
}
\value{
div element
}
\description{
In Dragula, the drag and drop areas are called "containers". This function
creates the individual draggable items and options in these containers.
}
