% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{dragulaZoneItems}
\alias{dragulaZoneItems}
\title{Create items for dragzones/dropzones}
\usage{
dragulaZoneItems(zone, type, items, multivalued = FALSE)
}
\arguments{
\item{zone}{Container zone type: either \code{drop} or \code{drag}}

\item{type}{Are these \code{options} or \code{presets}?}

\item{items}{List of item labels, with names corresponding to values}

\item{multivalued}{Multivalued? Only useful when \code{zone} is \code{drop}}
}
\value{
div element
}
\description{
In Dragula, the drag and drop areas are called "containers". This function
creates the individual draggable items and options in these containers.
}
