% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{dragZone}
\alias{dragZone}
\title{Create a dragzone container}
\usage{
dragZone(id, choices, ...)
}
\arguments{
\item{id}{The container id.}

\item{choices}{List of values to select from.}

\item{...}{Additional arguments passed along to tags$div, such as class}
}
\description{
Create a dragzone container
}
\examples{
dragZone("dragzone", choices = list(one = "One",
                                    two = "Two",
                                    three = "Three",
                                    four = "Four"))

}
\seealso{
\code{\link{dndselectr}}
}
