% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{dropZoneInput}
\alias{dropZoneInput}
\title{Create a Dragula dropzone input}
\usage{
dropZoneInput(inputId, choices, hidden = FALSE, placeholder = NULL,
  highlight = FALSE, multivalued = FALSE)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to acces the value.}

\item{choices}{List of acceptable values with their associated labels. Note that
the labels can be arbitrary HTML, as long as they are wrapped in a \code{tagList}.}

\item{hidden}{Should the selected items be hidden? This is useful to represent
a reactive or event trigger.}

\item{placeholder}{If hidden is true, insert placeholder text.}

\item{highlight}{Highlights the container on dragover. Useful when \code{hidden} is active.}

\item{multivalued}{Allow multiple items with the same value?}
}
\description{
Create a Dragula dropzone input
}
\examples{
dropZoneInput("dropzone", choices = list(one = "1",
                                         two = "2",
                                         three = "3",
                                         four = "4"))

}
\seealso{
\code{\link{dragulaSelectR}}
}
