% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndselect.R
\name{isMultivalued}
\alias{isMultivalued}
\title{Input is set of possible unique ids from multivalued inputs}
\usage{
isMultivalued(values)
}
\arguments{
\item{values}{Values returned by dropzone input}
}
\value{
logical
}
\description{
Input is set of possible unique ids from multivalued inputs
}
