% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragulaSelect.R
\name{multivalues}
\alias{multivalues}
\title{Convert multivalued inputs to values}
\usage{
multivalues(uniqueIds)
}
\arguments{
\item{uniqueIds}{Values returned by dropzone input}
}
\value{
Multivalues with unique id stripped away.
}
\description{
(Multivalued dropzones only) This will drop the added unique counter ids
to multivalued inputs.
}
